@@ -2,13 +2,12 @@
 
 import asyncio
 import fnmatch
-from typing import Callable, Iterable, Optional
+from typing import Callable, Iterable, Optional, AsyncGenerator
 from abc import ABC, abstractmethod
 from typing import List, Dict, Any
 
-from pydantic import functional_serializers
-
 from pywen.config.config import Config
+from pywen.ui.cli_console import CLIConsole
 from pywen.core.client import LLMClient
 from pywen.core.trajectory_recorder import TrajectoryRecorder
 from pywen.core.tool_registry import ToolRegistry
@@ -20,7 +19,7 @@ from pywen.tools.mcp_tool import MCPServerManager, sync_mcp_server_tools_into_re
 class BaseAgent(ABC):
     """Base class providing shared components for all agent implementations."""
     
-    def __init__(self, config: Config, cli_console=None):
+    def __init__(self, config: Config, cli_console: Optional[CLIConsole] =None):
         self.config = config
         self.cli_console = cli_console
         self.type = "BaseAgent"
@@ -54,7 +53,6 @@ class BaseAgent(ABC):
             for tool_name in failed_tools:
                 self.cli_console.print(f"Failed to register tool: {tool_name}", "yellow")
 
-
     async def setup_tools_mcp(self):
         """Setup tools based on agent configuration."""
         
@@ -75,7 +73,7 @@ class BaseAgent(ABC):
         self.cli_console = console
     
     @abstractmethod
-    async def run(self, user_message: str):
+    def run(self, user_message: str) -> AsyncGenerator[Dict[str, Any], None]:
         """Run the agent - must be implemented by subclasses."""
         pass
     
@@ -87,42 +85,31 @@ class BaseAgent(ABC):
     def reload_config(self):
         """重新加载配置"""
         try:
-            # 从正确的模块导入配置加载函数
-            from pywen.config.loader import load_config_from_file
-            
-            # 重新读取配置文件
-            new_config = load_config_from_file("pywen_config.json")
-            
-            # 保存旧的会话ID
+            from pywen.config.manager import ConfigManager 
+            cfg_mgr = ConfigManager("pywen_config.json")
+            new_config = cfg_mgr.load()
             old_session_id = getattr(self.config, 'session_id', None)
-            
-            # 更新配置
             self.config = new_config
-            
-            # 更新 max_iterations (如果 Agent 有这个属性)
             if hasattr(self, 'max_iterations'):
                 self.max_iterations = new_config.max_iterations
-            
-            # 恢复会话ID
             if old_session_id:
                 self.config.session_id = old_session_id
-            
-            # 重新初始化LLM客户端
             self.llm_client = LLMClient(new_config.model_config)
             
-            # 重新初始化task continuation checker (如果存在)
             if hasattr(self, 'task_continuation_checker'):
                 from pywen.agents.qwen.task_continuation_checker import TaskContinuationChecker
-                self.task_continuation_checker = TaskContinuationChecker(self.llm_client, new_config)
+                #TODO
+                self.task_continuation_checker = TaskContinuationChecker(self.llm_client)
             
             # 重建系统提示 (如果子类实现了该方法)
             if hasattr(self, '_build_system_prompt'):
                 self.system_prompt = self._build_system_prompt()
-            
-            self.cli_console.print(f"Config reloaded - Model: {new_config.model_config.model}, Max Steps: {new_config.max_iterations}")
+            if self.cli_console: 
+                self.cli_console.print(f"Config reloaded - Model: {new_config.model_config.model}, Max Steps: {new_config.max_iterations}")
             return True
         except Exception as e:
-            self.cli_console.print(f"Failed to reload config: {e}")
+            if self.cli_console: 
+                self.cli_console.print(f"Failed to reload config: {e}")
             return False
 
     def __make_include_predicate(self, patterns: Optional[Iterable[str]]) -> Optional[Callable[[str], bool]]:
