@@ -56,7 +56,7 @@ class QwenAgent(BaseAgent):
         self.loop_detector = AgentLoopDetectionService()
         
         # Initialize task continuation checker after llm_client is available
-        self.task_continuation_checker = TaskContinuationChecker(self.llm_client, config)
+        self.task_continuation_checker = TaskContinuationChecker(self.llm_client)
         
         # Conversation state
         self.turns: List[Turn] = []
@@ -110,7 +110,9 @@ class QwenAgent(BaseAgent):
         model_name = self.llm_client.utils_config.model_params.model
         # Get token limit from TokenLimits class
         max_tokens = TokenLimits.get_limit(ModelProvider.QWEN, model_name)
-        self.cli_console.set_max_context_tokens(max_tokens)
+        #TODO，从console剥离
+        if self.cli_console:
+            self.cli_console.set_max_context_tokens(max_tokens)
         
         # Reset task tracking for new user input
         self.original_user_task = user_message
@@ -129,10 +131,6 @@ class QwenAgent(BaseAgent):
             model=self.config.model_config.model,
             max_steps=self.max_iterations
         )
-        
-        # reset CLI tracking
-        if self.cli_console:
-            self.cli_console.reset_display_tracking()
         
         # Execute task with continuation logic in streaming mode
         current_message = user_message
@@ -875,6 +873,7 @@ Your core function is efficient and safe assistance. Balance extreme conciseness
                 if tool:
                     confirmation_details = await tool.get_confirmation_details(**tool_call.arguments)
                     if confirmation_details:  # 只有需要确认的工具才询问用户
+                        #TODO，从console剥离
                         confirmed = await self.cli_console.confirm_tool_call(tool_call, tool)
                         if not confirmed:
                                 # 用户拒绝，跳过这个工具
