@@ -1,13 +1,13 @@
 """内存管理命令实现"""
 
 from rich.panel import Panel
-from rich.console import Console
+from rich import get_console
 from .base_command import BaseCommand
 
 class MemoryCommand(BaseCommand):
     def __init__(self):
         super().__init__("memory", "Commands for interacting with memory.")
-        self.console = Console()
+        self.console = get_console()
     
     async def execute(self, context, args: str) -> bool:
         """处理内存相关命令"""
